vm::compile
